% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mosaic.R
\name{animate_conus_mosaic}
\alias{animate_conus_mosaic}
\title{Create an animated weather radar mosaic for the the conterminus U.S.}
\usage{
animate_conus_mosaic(size = c("small", "large"))
}
\arguments{
\item{size}{one of "\code{small}" or "\code{large}". The large mosaic takes a bit to download.}
}
\description{
Create an animated weather radar mosaic for the the conterminus U.S.
}
