% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animate.R
\name{animate_radar}
\alias{animate_radar}
\title{Create an animated weather image from a NOAA station}
\usage{
animate_radar(station = "GYX")
}
\arguments{
\item{station}{the abbreviated station name. See \link{stations}.}
}
\description{
Create an animated weather image from a NOAA station
}
