% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latest_radar.R
\name{latest_radar}
\alias{latest_radar}
\title{Read latest NWS regional or ConUS radar mosaics as a \code{stars} object}
\usage{
latest_radar(
  regional_mosaic = c("alaska", "centgrtlakes", "hawaii", "latest", "northeast",
    "northrockies", "pacnorthwest", "pacsouthwest", "southeast", "southmissvly",
    "southplains", "southrockies", "uppermissvly"),
  quiet = TRUE
)
}
\arguments{
\item{regional_mosaic}{one of the supported mosaics from \href{https://radar.weather.gov/ridge/}{the official NWS list}:
\itemize{
\item "\code{alaska}": Alaska
\item "\code{centgrtlakes}": Central Great Lakes
\item "\code{hawaii}": Hawaii
\item "\code{latest}": ConUS
\item "\code{northeast}": Northeast
\item "\code{northrockies}": Northern Rockies
\item "\code{pacnorthwest}": Pacific Northwest
\item "\code{pacsouthwest}": Pacific Southwest
\item "\code{southeast}": Southeast
\item "\code{southmissvly}": Southern Mississippi Valley
\item "\code{southplains}": Southern Plains
\item "\code{southrockies}": Southern Rockies
\item "\code{uppermissvly}": Upper Mississippi Valley
}}

\item{quiet}{passed on to \code{\link[utils:download.file]{utils::download.file()}}}
}
\value{
\code{stars} object with the raster composite with a CRS of EPSG:4326
}
\description{
Read latest NWS regional or ConUS radar mosaics as a \code{stars} object
}
