% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latest_ridge.R
\name{latest_ridge}
\alias{latest_ridge}
\title{Read latest NWS ridge as a stars object}
\usage{
latest_ridge(
  station = "GYX",
  ridge = c("base", "storm", "hour", "composite", "storm_total", "base_extended"),
  quiet = TRUE
)
}
\arguments{
\item{station}{the abbreviated station name. See \link{stations}.}

\item{ridge}{image type
\itemize{
\item "\code{base}" for Base Reflectivity (out to 124 nm). \strong{The default}.
\item "\code{storm}" for Storm Relative Motion
\item "\code{hour}" for One-Hour Precipitation
\item "\code{composite}" for Composite Reflectivity
\item "\code{storm_total}" for Storm Total Precipitation
\item "\code{base_extended}" for Base Reflectivity (out to 248 nmi)
}}

\item{quiet}{passed on to \code{\link[utils:download.file]{utils::download.file()}}}
}
\value{
\code{stars} object with the raster composite with a CRS of EPSG:4326
}
\description{
Read latest NWS ridge as a stars object
}
