% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rradar-package.R
\docType{package}
\name{rradar}
\alias{rradar}
\alias{rradar-package}
\title{...}
\description{
A good description goes here otherwise CRAN checks fail.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://gitlab.com/hrbrmstr/rradar}
  \item Report bugs at \url{https://gitlab.com/hrbrmstr/rradar/issues}
}

}
\author{
Bob Rudis (bob@rud.is)
}
\keyword{internal}
