% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stations-df.R
\docType{data}
\name{stations}
\alias{stations}
\title{NOAA U.S. Radar Stations}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 179 rows and 4 columns.}
\usage{
stations
}
\description{
NOAA U.S. Radar Stations
}
\keyword{datasets}
