% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rradar-package.R
\docType{package}
\name{rradar}
\alias{rradar}
\alias{rradar-package}
\title{Animate NOAA NWS N0R Radar Images by Station Id}
\description{
NOAA NWS has an array of National Doppler Radar Sites. Tools are provided to
to help you locate stations and create an animated composite image of recent radar
images.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://gitlab.com/hrbrmstr/rradar}
  \item Report bugs at \url{https://gitlab.com/hrbrmstr/rradar/issues}
}

}
\author{
Bob Rudis (bob@rud.is)
}
\keyword{internal}
